added baseViewmodel and refactored it
