+++ app/src/main/java/com/one4ll/xplayer/ui/stream/StreamFragment.kt
@@ -19,6 +19,7 @@ import com.one4ll.xplayer.adapter.StreamsRecyclerViewAdapter
 import com.one4ll.xplayer.database.MediaDatabase
 import com.one4ll.xplayer.databinding.FragmentStreamBinding
 import com.one4ll.xplayer.helpers.VIDEO_PATH
+import com.one4ll.xplayer.helpers.baseViewModel
 import com.one4ll.xplayer.models.Streams
 import kotlinx.android.synthetic.main.fragment_stream.*
 import kotlinx.android.synthetic.main.fragment_stream.view.*
@@ -37,7 +38,6 @@ private const val TAG = "streamFragment"
 @ExperimentalCoroutinesApi
 class StreamFragment : Fragment() {
     private val adapter: StreamsRecyclerViewAdapter by lazy { StreamsRecyclerViewAdapter() }
-    private val viewModel: StreamViewModel by viewModels()
     private lateinit var binding: FragmentStreamBinding
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?,
@@ -51,7 +51,7 @@ class StreamFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         setRecycleView()
         lifecycleScope.launch {
-            viewModel.streamsList.collect {
+            baseViewModel.streamsList.collect {
                 adapter.setNotes(it)
             }
         }
@@ -61,7 +61,7 @@ class StreamFragment : Fragment() {
                 val intent = Intent(binding.root.context, MainActivity::class.java)
                 intent.putExtra(VIDEO_PATH, url.toString())
                 lifecycleScope.launch {
-                    insertStreamsIntoDatabase(viewModel.db, Streams(url.toString(), System.currentTimeMillis()))
+                    insertStreamsIntoDatabase(baseViewModel.db, Streams(url.toString(), System.currentTimeMillis()))
                 }
                 Log.d(TAG, "onCreateView: stream url path $url")
                 binding.root.context.startActivity(intent)
@@ -78,7 +78,7 @@ class StreamFragment : Fragment() {
             override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {
                 Log.d(TAG, "onSwiped: direction $direction")
                 lifecycleScope.launch {
-                    viewModel.db.streamsDao().removeById(adapter.getStreamAtPosition(viewHolder.adapterPosition).id!!)
+                    baseViewModel.db.streamsDao().removeById(adapter.getStreamAtPosition(viewHolder.adapterPosition).id!!)
                     adapter.notifyDataSetChanged()
                 }
             }
